pyproject.toml
[project]
name = "tabular_pipeline"
version = "0.1.0"
description = "A modern, reusable pipeline for tabular ML projects."
authors = [
    {name = "Your Name", email = "your@email.com"},
]
dependencies = [
    "catboost>=1.2.8",
    "hydra-core>=1.3.2",
    "mlflow>=3.1.4",
    "pandera>=0.26.1",
    "scikit-learn>=1.6.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
    "dbt-duckdb>=1.7.0",
    "polars>=0.20.0",
    "hydra-core>=1.3.0",
    "mlflow>=2.12.0",
    "ray[tune]>=2.9.0",
    "catboost>=1.2.0",
    "torch>=2.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "marimo>=0.4.0",
]
